refactor: move component
