+++ front/src/components/Gainburua/AtalaName.ts
@@ -1,0 +1,1 @@
+export type AtalaName = 'hasiera' | 'kafetegia';
