+++ front/src/components/Gainburua/components/DesktopNabigazioa/index.ts
@@ -1,0 +1,1 @@
+export { DesktopNabigazioa } from './DesktopNabigazioa';
